requirements.txt
PyQt5==5.15.11
PyInstaller==5.10.1
python-dotenv==1.0.1